# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proector API",
        "description": "Backend API documentation for Proector mobile and admin platforms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base path"
        },
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/api/admin/teams": {
            "get": {
                "tags": [
                    "Admin Teams"
                ],
                "summary": "Get all security teams",
                "description": "Returns all security teams with personnel and currently active bookings.",
                "operationId": "764f07b4be9872117aef38607cdfb26b",
                "responses": {
                    "200": {
                        "description": "Teams list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Teams"
                ],
                "summary": "Create security team",
                "description": "Creates a new security team with service type and team size.",
                "operationId": "225d7a32d9d6b773a1a5a9e183e4c42d",
                "responses": {
                    "201": {
                        "description": "Team created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/vehicles": {
            "get": {
                "tags": [
                    "Admin Vehicles"
                ],
                "summary": "Get all vehicles",
                "description": "Returns all vehicles with availability and usage metadata.",
                "operationId": "18c0090655662b5e48e91b9f82637cfb",
                "responses": {
                    "200": {
                        "description": "Vehicles list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Vehicles"
                ],
                "summary": "Create vehicle",
                "description": "Creates a vehicle record that can be assigned to bookings.",
                "operationId": "ffffef66b3510d35075d5128bc45c0b7",
                "responses": {
                    "201": {
                        "description": "Vehicle created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/payments/webhook": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Handle payment webhook",
                "description": "Processes payment provider webhook events with idempotency protection.",
                "operationId": "45a347f7120201c07889a9a050a43164",
                "responses": {
                    "200": {
                        "description": "Webhook processed"
                    },
                    "400": {
                        "description": "Invalid webhook signature"
                    }
                }
            }
        },
        "/api/security/login": {
            "post": {
                "tags": [
                    "Security Personnel Auth"
                ],
                "summary": "Login security personnel",
                "description": "Authenticates security personnel with username/password and returns token.",
                "operationId": "2df18ec2e71faf843c0a52ddfee6697c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    }
                }
            }
        },
        "/api/security/logout": {
            "post": {
                "tags": [
                    "Security Personnel Auth"
                ],
                "summary": "Logout security personnel",
                "description": "Revokes current security personnel access token.",
                "operationId": "37ad16087a83fcdec6acff9e089f18b9",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/security/orders/{id}/messages": {
            "get": {
                "tags": [
                    "Security Personnel Chat"
                ],
                "summary": "Get messages for order",
                "description": "Returns chat thread for a security-assigned booking.",
                "operationId": "24523dd23fc92aac1cf358f1764306ab",
                "responses": {
                    "200": {
                        "description": "Messages"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Security Personnel Chat"
                ],
                "summary": "Send message",
                "description": "Sends chat message from security personnel to client and broadcasts realtime event.",
                "operationId": "863019cdc69494fa8b0e930969037dc0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string"
                                    },
                                    "message_type": {
                                        "type": "string",
                                        "default": "text",
                                        "enum": [
                                            "text",
                                            "sos",
                                            "quick_response"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message sent"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Bearer {token} from login or OTP verification endpoints.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}